docs/requirements.txt
# Copyright (c) Ansible Project
# GNU General Public License v3.0+ (see LICENSES/GPL-3.0-or-later.txt or https://www.gnu.org/licenses/gpl-3.0.txt)
# SPDX-License-Identifier: GPL-3.0-or-later

ansible-core>=2.13,<2.15
antsibull-docs >= 1.0.0, < 2.0.0
ansible-pygments
sphinx
sphinx-ansible-theme >= 0.9.0
pyyaml
